refactored sorting function, now it also works for group results
